Move the registration tests next to their sources.
1. Group `_apply_transform` and `_learn_transform` into a `_registration` package.  This is because there is common code to the two (basically `TransformsList`).
2. Move tests into each package.  `test_registration.py` is broken up into multiple files, each tied to the package they reside in.
3. Ninja'd in the removal of `starfish/spots/test/__init__.py`.  Not even sure why that's there.
