+++ starfish/image/__init__.py
@@ -1,4 +1,4 @@
-from ._apply_transform import ApplyTransform
 from ._filter import Filter
-from ._learn_transform import LearnTransform
+from ._registration._apply_transform import ApplyTransform
+from ._registration._learn_transform import LearnTransform
 from ._segmentation import Segmentation
