+++ starfish/image/_registration/_learn_transform/_base.py
@@ -2,7 +2,7 @@ from abc import abstractmethod
 from typing import Type
 
 
-from starfish.image._learn_transform.transforms_list import TransformsList
+from starfish.image._registration.transforms_list import TransformsList
 from starfish.imagestack.imagestack import ImageStack
 from starfish.pipeline.algorithmbase import AlgorithmBase
 from starfish.pipeline.pipelinecomponent import PipelineComponent
